Working on api

+++ src/main/java/com/angelina/myWallet/controllers/ExpenseController.java
@@ -3,6 +3,7 @@ package com.angelina.myWallet.controllers;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -10,14 +11,7 @@ import com.angelina.myWallet.model.Expense;
 import com.angelina.myWallet.repository.ExpenseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api")
@@ -27,19 +21,24 @@ public class ExpenseController {
     private ExpenseRepository expenseRepository;
 
     @GetMapping("/expenses")
-    List<Expense> getExpenses(){
+    public List<Expense> getExpenses() {
         return expenseRepository.findAll();
     }
 
     @DeleteMapping("/expenses/{id}")
-    ResponseEntity<?>  deleteExpense(@PathVariable Long id){
-        expenseRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<?> deleteExpense(@PathVariable Long id) {
+        Optional<Expense> expense = expenseRepository.findById(id);
+        if (expense.isPresent()) {
+            expenseRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping("/expenses")
-    ResponseEntity<Expense> createExpense(@Valid @RequestBody Expense expense) throws URISyntaxException{
-        Expense result= expenseRepository.save(expense);
-        return ResponseEntity.created(new URI("/api/expenses" + result.getId())).body(result);
+    public ResponseEntity<Expense> createExpense(@Valid @RequestBody Expense expense) throws URISyntaxException {
+        Expense result = expenseRepository.save(expense);
+        return ResponseEntity.created(new URI("/api/expenses/" + result.getId())).body(result);
     }
 }
